requirements.txt
streamlit==1.28.0
google-generativeai==0.3.0
pdfplumber==0.9.0
PyMuPDF==1.23.0
weasyprint==60.0
psycopg2-binary==2.9.7
streamlit-authenticator==0.2.3
streamlit-option-menu==0.3.6
python-dotenv==1.0.0
Pillow==10.0.0
pandas==2.1.0
bcrypt==4.0.1
